+++ src/transports/smart.c
@@ -24,7 +24,7 @@ static int git_smart__recv_cb(gitno_buffer *buf)
 	buf->offset += bytes_read;
 
 	if (t->packetsize_cb)
-		t->packetsize_cb((int)bytes_read, t->packetsize_payload);
+		t->packetsize_cb(bytes_read, t->packetsize_payload);
 
 	return (int)(buf->offset - old_len);
 }
